chore(linter): fix linter issues (#1911)
* linter over app/

* auto fix over x/

* more fixes over x/

* fixinig ibctesting

* fixed tests after adding error check on dymns

* fixed casting checks

* fixed int overflow on streamer

* ignore false positive on ibctesting

* fixed delayack

* fixed eibc

* fixed lockup

* more fixes

* fixed non tests linter

* ignoring conversion linter on test files

* refactor test assertions to include type checks for expected values

* disable failed protobuf check

* increased timeout

+++ x/delayedack/types/params.go
@@ -13,11 +13,11 @@ const (
 )
 
 // NewParams creates a new Params instance
-func NewParams(epochIdentifier string, bridgingFee math.LegacyDec, deletePacketsEpochLimit int) Params {
+func NewParams(epochIdentifier string, bridgingFee math.LegacyDec, deletePacketsEpochLimit int32) Params {
 	return Params{
 		EpochIdentifier:         epochIdentifier,
 		BridgingFee:             bridgingFee,
-		DeletePacketsEpochLimit: int32(deletePacketsEpochLimit),
+		DeletePacketsEpochLimit: deletePacketsEpochLimit,
 	}
 }
